pyproject.toml
[project]
name = "vector-database"
version = "0.1.0"
description = ""
authors = [
    {name = "ayalamanu",email = "manu.ayala05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "qdrant-client (>=1.14.3,<2.0.0)",
    "numpy (==2.2)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.0,<0.117.0)",
    "starlette < 0.47.0",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false